added SpectrumLike DispersionSpectrumLike and updated OGIPLike
